Move windows helper functions to header file
